isolate redis testing instance
